use the layer/charm options for minimum channel support in the release jobs
