+++ jobs/build-charms/charms.py
@@ -27,11 +27,11 @@ from cilib.git import default_gh_branch
 from cilib.enums import SNAP_K8S_TRACK_MAP, K8S_SERIES_MAP, K8S_CHARM_SUPPORT_ARCHES
 from cilib.service.aws import Store
 from cilib.run import cmd_ok, capture, script
+from cilib.version import ChannelRange, Release, RISKS
 from datetime import datetime
-from dataclasses import dataclass
 from enum import Enum
 from types import SimpleNamespace
-from typing import List, Mapping, Optional, Union
+from typing import List, Mapping, Optional
 
 from pathos.threading import ThreadPool
 from pprint import pformat
@@ -42,49 +42,6 @@ import yaml
 import json
 import requests
 import re
-
-
-RISKS = ["stable", "candidate", "beta", "edge"]
-
-
-@dataclass
-class Release:
-    """Representation of Charm or Snap Release version."""
-
-    major: int
-    minor: int
-    risk: Optional[str]
-
-    def __str__(self):
-        risk = ""
-        if self.risk:
-            risk = f"/{self.risk}" if self.risk.lower() in RISKS else ""
-        return f"{self.major}.{self.minor}{risk}"
-
-    def _as_cmp(self):
-        return (
-            self.major,
-            self.minor,
-            RISKS[::-1].index(self.risk.lower()) + 1 if self.risk else 0,
-        )
-
-    @classmethod
-    def mk(cls, rel: str) -> "Release":
-        has_risk = rel.split("/")
-        if len(has_risk) == 2:
-            track, risk = has_risk
-        else:
-            track, risk = has_risk[0], None
-        return cls(*map(int, track.split(".")), risk)
-
-    def __eq__(self, other: "Release") -> bool:
-        return self._as_cmp() == other._as_cmp()
-
-    def __gt__(self, other: "Release") -> bool:
-        return self._as_cmp() > other._as_cmp()
-
-    def __lt__(self, other: "Release") -> bool:
-        return self._as_cmp() < other._as_cmp()
 
 
 def matched_numerical_channel(
@@ -103,44 +60,6 @@ def matched_numerical_channel(
             chan = f"{release}/{risk}"
             if chan in tracks:
                 return chan
-
-
-@dataclass
-class ChannelRange:
-    """Determine if channel is within a channel range.
-
-    Usage:
-        assert "latest/edge" in ChannelRange("1.18", "1.25/stable") # latest/<anything> is ignored
-        assert "1.24/edge" in ChannelRange("1.18", "1.25/stable")   # within bounds inclusively
-        assert "1.24/edge" in ChannelRange("1.18", None)            # No upper bound
-        assert "1.24/edge" in ChannelRange(None, "1.25/stable")     # No lower bound
-        assert "1.24/edge" in ChannelRange(None, None)              # No bounds
-    """
-
-    _min: Optional[str]
-    _max: Optional[str]
-
-    @property
-    def min(self) -> Optional[Release]:
-        """Release object representing the minimum."""
-        return self._min and Release.mk(self._min)
-
-    @property
-    def max(self) -> Optional[Release]:
-        """Release object representing the maximum."""
-        return self._max and Release.mk(self._max)
-
-    def __contains__(self, other: Union[str, Release]) -> bool:
-        """Implements comparitor."""
-        if isinstance(other, str) and other.startswith("latest"):
-            return True
-        if not isinstance(other, Release):
-            other = Release.mk(str(other))
-        if self.min and other < self.min:
-            return False
-        if self.max and other > self.max:
-            return False
-        return True
 
 
 def generate_manifest(reactive_charm, archs):
